Move BinderTransport's inner classes to the top level
